+++ client/src/pages/Chart.tsx
@@ -1,0 +1,11 @@
+import Container from "../components/Container";
+import PieChart from "../components/PieChart";
+
+export default function Chart() {
+  return (
+    <Container>
+      <h1>Chart</h1>
+      <PieChart />
+    </Container>
+  );
+}
